# perf: add optional support for keccak-asm