--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,7 +10,7 @@ authors = ["Alloy Contributors"]
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/alloy-rs/core"
 repository = "https://github.com/alloy-rs/core"
-exclude = []
+exclude = ["tests"]
 
 [workspace.metadata.docs.rs]
 all-features = true