--- a/crates/primitives/Cargo.toml
+++ b/crates/primitives/Cargo.toml
@@ -96,12 +96,6 @@ arbitrary = [
 hex-compat = ["hex/hex"]
 
 [[bench]]
-name = "address"
-path = "benches/address.rs"
-harness = false
-required-features = ["rand"]
-
-[[bench]]
 name = "primitives"
 path = "benches/primitives.rs"
 harness = false