--- a/crates/primitives/benches/primitives.rs
+++ b/crates/primitives/benches/primitives.rs
@@ -1,6 +1,4 @@
-#![allow(unused)]
-
-use alloy_primitives::Address;
+use alloy_primitives::{keccak256, Address, B256};
 use criterion::{criterion_group, criterion_main, Criterion};
 use std::hint::black_box;
 
@@ -15,11 +13,10 @@ fn primitives(c: &mut Criterion) {
         })
     });
     g.bench_function("keccak256/32", |b| {
-        let mut out = alloy_primitives::B256::random();
+        let mut out = B256::random();
         b.iter(|| {
-            #[allow(clippy::needless_borrows_for_generic_args)]
             for _ in 0..10 {
-                out = alloy_primitives::keccak256(&out);
+                out = keccak256(out);
             }
             black_box(&out);
         });