--- a/crates/primitives/src/utils/mod.rs
+++ b/crates/primitives/src/utils/mod.rs
@@ -62,7 +62,7 @@ pub fn keccak256<T: AsRef<[u8]>>(bytes: T) -> B256 {
         let mut output = MaybeUninit::<B256>::uninit();
 
         cfg_if::cfg_if! {
-            if #[cfg(all(feature = "native-keccak", not(feature = "tiny-keccak")))] {
+            if #[cfg(all(feature = "native-keccak", not(feature = "tiny-keccak"), not(miri)))] {
                 #[link(wasm_import_module = "vm_hooks")]
                 extern "C" {
                     /// When targeting VMs with native keccak hooks, the `native-keccak` feature
@@ -84,7 +84,7 @@ pub fn keccak256<T: AsRef<[u8]>>(bytes: T) -> B256 {
 
                 // SAFETY: The output is 32-bytes, and the input comes from a slice.
                 unsafe { native_keccak256(bytes.as_ptr(), bytes.len(), output.as_mut_ptr().cast()) };
-            } else if #[cfg(feature = "asm-keccak")] {
+            } else if #[cfg(all(feature = "asm-keccak", not(miri)))] {
                 use keccak_asm::{digest::Digest, Keccak256};
 
                 let mut hasher = Keccak256::new();