<!--
Thank you for your Pull Request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide: https://github.com/alloy-rs/core/blob/main/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options ** -->

## Motivation

Closes #414 
See https://github.com/DaniPopes/keccak-asm, https://github.com/DaniPopes/bench-keccak256

<!--
Explain the context and why you're making that change. What is the problem
you're trying to solve? In some cases there is not a problem and this can be
thought of as being the motivation for your change.
-->

## Solution

Add `"asm-keccak"` feature which overrides the tiny-keccak implementation with [`keccak-asm`](https://crates.io/crates/keccak-asm)

<!--
Summarize the solution and provide any necessary context needed to understand
the code change.
-->

## Benches

https://github.com/DaniPopes/bench-keccak256

Local:

RUSTFLAGS=""
2.60-2.75 µs -> 2.10-2.20 µs

```
primitives/keccak256/32 time:   [2.1212 µs 2.1285 µs 2.1373 µs]                                     
                        change: [-21.731% -21.336% -20.849%] (p = 0.00 < 0.05)
                        Performance has improved.
```

RUSTFLAGS="-Ctarget-cpu=native" (AVX512VL)
2.08-2.20 µs -> 2.04-2.10 µs

```
primitives/keccak256/32 time:   [2.0734 µs 2.0747 µs 2.0761 µs]                                     
                        change: [-5.1237% -5.0213% -4.8902%] (p = 0.00 < 0.05)
                        Performance has improved.
```

## PR Checklist

- [ ] Added Tests
- [ ] Added Documentation
- [ ] Breaking changes